pages comp
