+++ frontend/src/app/page.js
@@ -1,4 +1,4 @@
-import Homepage from '@/components/pages/Homepage';
+import Homepage from "@/components/pages/Homepage";
 
 export default function Home() {
